+++ Photon.Agent/ApiHandlers/LogHandler.cs
