+++ Photon.Agent/ApiHandlers/Session/ActiveHandler.cs
