+++ Photon.Agent/ApiHandlers/Session/CancelHandler.cs
