+++ Photon.Agent/ApiHandlers/Session/OutputStreamHandler.cs
