+++ Photon.Agent/ApiHandlers/Variable/DeleteHandler.cs
