+++ Photon.Agent/ApiHandlers/Variable/JsonHandler.cs
