+++ Photon.Agent/PhotonAgent.csproj
@@ -90,17 +90,17 @@
       <DependentUpon>AgentService.cs</DependentUpon>
     </Compile>
     <Compile Include="Commands\RootCommands.cs" />
-    <Compile Include="HttpHandlers\Api\LogHandler.cs" />
-    <Compile Include="HttpHandlers\Api\Session\ActiveHandler.cs" />
-    <Compile Include="HttpHandlers\Api\Session\CancelHandler.cs" />
-    <Compile Include="HttpHandlers\Api\Session\OutputStreamHandler.cs" />
-    <Compile Include="HttpHandlers\Api\Variable\DeleteHandler.cs" />
-    <Compile Include="HttpHandlers\Api\Variable\JsonHandler.cs" />
-    <Compile Include="HttpHandlers\ConfigurationHandler.cs" />
-    <Compile Include="HttpHandlers\Session\SessionDetailsHandler.cs" />
-    <Compile Include="HttpHandlers\Session\SessionIndexHandler.cs" />
-    <Compile Include="HttpHandlers\VariableSet\VariableSetEditJsonHandler.cs" />
-    <Compile Include="HttpHandlers\VariableSet\VariableSetIndexHandler.cs" />
+    <Compile Include="ApiHandlers\LogHandler.cs" />
+    <Compile Include="ApiHandlers\Session\ActiveHandler.cs" />
+    <Compile Include="ApiHandlers\Session\CancelHandler.cs" />
+    <Compile Include="ApiHandlers\Session\OutputStreamHandler.cs" />
+    <Compile Include="ApiHandlers\Variable\DeleteHandler.cs" />
+    <Compile Include="ApiHandlers\Variable\JsonHandler.cs" />
+    <Compile Include="ViewHandlers\ConfigurationHandler.cs" />
+    <Compile Include="ViewHandlers\Session\SessionDetailsHandler.cs" />
+    <Compile Include="ViewHandlers\Session\SessionIndexHandler.cs" />
+    <Compile Include="ViewHandlers\VariableSet\VariableSetEditJsonHandler.cs" />
+    <Compile Include="ViewHandlers\VariableSet\VariableSetIndexHandler.cs" />
     <Compile Include="Internal\Git\CmdCheckout.cs" />
     <Compile Include="Internal\Git\ICheckout.cs" />
     <Compile Include="Internal\Git\LibCheckout.cs" />
@@ -130,7 +130,7 @@
     <Compile Include="MessageHandlers\TaskRunProcessor.cs" />
     <Compile Include="MessageHandlers\BuildSessionReleaseProcessor.cs" />
     <Compile Include="MessageHandlers\BuildSessionBeginProcessor.cs" />
-    <Compile Include="HttpHandlers\IndexHandler.cs" />
+    <Compile Include="ViewHandlers\IndexHandler.cs" />
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="ViewModels\IndexVM.cs" />
