+++ Photon.Agent/ViewHandlers/ConfigurationHandler.cs
