+++ Photon.Agent/ViewHandlers/IndexHandler.cs
