+++ Photon.Agent/ViewHandlers/Session/SessionDetailsHandler.cs
