+++ Photon.Agent/ViewHandlers/Session/SessionIndexHandler.cs
