+++ Photon.Agent/ViewHandlers/VariableSet/VariableSetEditJsonHandler.cs
