+++ Photon.Agent/ViewHandlers/VariableSet/VariableSetIndexHandler.cs
