--- Photon.Framework/Server/IAgentSessionHandle.cs
@@ -1,16 +1,0 @@
-﻿using Photon.Framework.Tasks;
-using System;
-using System.Threading.Tasks;
-
-namespace Photon.Framework.Server
-{
-    public interface IAgentSessionHandle : IDisposable
-    {
-        TaskRunnerManager Tasks {get;}
-        string AgentSessionId {get;}
-
-        Task BeginAsync();
-        Task ReleaseAsync();
-        Task<TaskResult> RunTaskAsync(string taskName);
-    }
-}
