+++ Photon.Server/PhotonServer.csproj
@@ -117,7 +117,7 @@
     <Compile Include="Internal\Builds\BuildData.cs" />
     <Compile Include="Internal\Builds\BuildDataManager.cs" />
     <Compile Include="Internal\ClientHandshake.cs" />
-    <Compile Include="Internal\Projects\ProjectManager2.cs" />
+    <Compile Include="Internal\Projects\ProjectManager.cs" />
     <Compile Include="Internal\ServerAgents\ServerAgentManager.cs" />
     <Compile Include="Internal\GitHub\HookEventHandler.cs" />
     <Compile Include="Internal\JsonDynamicDocument.cs" />
